fix(evs): move resource evs to dir /huaweicloud/service/evc
